+++ pkg/BUILD.bazel
@@ -64,7 +64,6 @@ ALL_TESTS = [
     "//pkg/ccl/multitenantccl/tenantcostclient:tenantcostclient_test",
     "//pkg/ccl/multitenantccl/tenantcostserver/tenanttokenbucket:tenanttokenbucket_test",
     "//pkg/ccl/multitenantccl/tenantcostserver:tenantcostserver_test",
-    "//pkg/ccl/oidcccl:oidcccl_test",
     "//pkg/ccl/schemachangerccl/sctestbackupccl:sctestbackupccl_test",
     "//pkg/ccl/schemachangerccl/sctestbackupmrccl:sctestbackupmrccl_test",
     "//pkg/ccl/schemachangerccl:schemachangerccl_test",
@@ -377,6 +376,7 @@ ALL_TESTS = [
     "//pkg/security/dyncert:dyncert_test",
     "//pkg/security/jwtauth:jwtauth_test",
     "//pkg/security/ldapauth:ldapauth_test",
+    "//pkg/security/oidcauth:oidcauth_test",
     "//pkg/security/password:password_test",
     "//pkg/security/provisioning:provisioning_test",
     "//pkg/security/sessionrevival:sessionrevival_test",
@@ -1044,8 +1044,6 @@ GO_TARGETS = [
     "//pkg/ccl/multitenantccl/tenantcostserver:tenantcostserver",
     "//pkg/ccl/multitenantccl/tenantcostserver:tenantcostserver_test",
     "//pkg/ccl/multitenantccl:multitenantccl",
-    "//pkg/ccl/oidcccl:oidcccl",
-    "//pkg/ccl/oidcccl:oidcccl_test",
     "//pkg/ccl/schemachangerccl/sctestbackupccl:sctestbackupccl_test",
     "//pkg/ccl/schemachangerccl/sctestbackupmrccl:sctestbackupmrccl_test",
     "//pkg/ccl/schemachangerccl:schemachangerccl",
@@ -1958,6 +1956,8 @@ GO_TARGETS = [
     "//pkg/security/jwthelper:jwthelper",
     "//pkg/security/ldapauth:ldapauth",
     "//pkg/security/ldapauth:ldapauth_test",
+    "//pkg/security/oidcauth:oidcauth",
+    "//pkg/security/oidcauth:oidcauth_test",
     "//pkg/security/password:password",
     "//pkg/security/password:password_test",
     "//pkg/security/pprompt:pprompt",
